fix: Fixed bug where incorrect type would be passed and would cause id to get lost

+++ src/Renderer/Models/LineModel.cpp
@@ -61,7 +61,7 @@ const float &LineModel::GetWidth()
     return width;
 }
 
-void LineModel::SetModelId(long long int id)
+void LineModel::SetModelId(size_t id)
 {
     this->id = id;
 }
